+++ app/core/api/views.py
@@ -1,5 +1,4 @@
 from django.contrib.auth import get_user_model
-from django.http import Http404
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiExample
 from drf_spectacular.utils import OpenApiParameter
@@ -11,7 +10,6 @@ from rest_framework.generics import GenericAPIView
 from rest_framework.mixins import RetrieveModelMixin
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from django.contrib.auth.models import Group
 
 from ..models import Affiliate
 from ..models import Affiliation
@@ -51,10 +49,8 @@ from .serializers import StackElementTypeSerializer
 from .serializers import UrlTypeSerializer
 from .serializers import UserPermissionSerializer
 from .serializers import UserSerializer
-from .serializers import UserAppSerializer
 from .serializers import UserStatusTypeSerializer
-from .helpers import filter_user_queryset
-from rest_framework.exceptions import PermissionDenied
+
 
 
 
@@ -169,28 +165,6 @@ class UserViewSet(viewsets.ModelViewSet):
     ),
     retrieve=extend_schema(description="Return the given user"),
 )
-class UserAppKbViewSet(viewsets.ModelViewSet):
-    permission_classes = [IsAuthenticated]
-    serializer_class = UserAppSerializer
-    lookup_field = "uuid"
-    calling_app = "kb" # using variable to make extending to other apps easier
-
-    def get_queryset(self):
-        """
-        Optionally filter users by an 'email' and/or 'username' query paramerter in the URL
-        """
-        user = self.request.user
-        permission_name = f"get_api_user_app_{self.calling_app}"
-        if not user.has_perm(permission_name):
-            raise PermissionDenied(f"You don't have privilege to view users for calling app {calling_app}.")
-        groups = Group.objects.filter(name__startswith=f"{self.calling_app}_")
-
-        # Filter users who are in any of these groups
-        queryset = get_user_model.objects.filter(groups__in=groups).distinct()
-        queryset = filter_user_querset()
-        return queryset
-
-
 @extend_schema_view(
     list=extend_schema(description="Return a list of all the projects"),
     create=extend_schema(description="Create a new project"),
